Add authentication and user games management functionality
- Implemented authController for user registration, login, and JWT token generation.
- Created userGamesController for managing user-specific game ratings and statuses.
- Added routes for authentication and user games management.
- Introduced a migration script to update the rating column type in the database.
- Enhanced frontend with delete functionality for user games and improved UI elements.

+++ backend/src/db/init.ts
@@ -46,7 +46,7 @@ async function initializeDatabase() {
         id SERIAL PRIMARY KEY,
         user_id INTEGER REFERENCES users(id),
         game_id INTEGER REFERENCES games(id),
-        rating INTEGER,
+        rating DECIMAL(3,1),
         status TEXT,
         created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         UNIQUE(user_id, game_id)
